+++ CHANGELOG.md
@@ -48,6 +48,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Show only owned reports and dashboards if not admin
 - Fixed usage of an order column when that column wasn't appearing in the
   SELECT clause
+- Error handler sent by server to be displayed on the client-side
 
 ### Dependencies